basic home and about pages, organized files and directories
