+++ src/containers/About.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function About() {
+  return (
+    <div>
+      About
+    </div>
+  );
+}
+
+export default About;
